+++ lib/src/controller/interface/kakao_map_controller_platform_interface.dart
@@ -1,4 +1,10 @@
-part of '../kakao_map_controller.dart';
+import 'dart:async' show StreamController;
+
+import 'package:flutter/foundation.dart';
+import 'package:kakao_maps_flutter/src/data/data.dart';
+import 'package:kakao_maps_flutter/src/method_call/kakao_map_method_call.dart';
+import 'package:kakao_maps_flutter/src/method_call/kakao_map_web_method_call.dart';
+import 'package:plugin_platform_interface/plugin_platform_interface.dart';
 
 /// Platform interface for Kakao Map controller functionality.
 ///
@@ -9,21 +15,6 @@ abstract class KakaoMapControllerPlatform extends PlatformInterface {
   KakaoMapControllerPlatform() : super(token: _token);
 
   static final Object _token = Object();
-
-  /// The current platform instance.
-  static KakaoMapControllerPlatform get instance => _instance;
-
-  static KakaoMapControllerPlatform _instance =
-      MethodChannelKakaoMapController.instance;
-
-  /// Sets the platform instance.
-  ///
-  /// Platform implementations should set this to register themselves
-  /// as the active platform implementation.
-  static set instance(KakaoMapControllerPlatform instance) {
-    PlatformInterface.verifyToken(instance, _token);
-    _instance = instance;
-  }
 
   /// Stream controller for label click events.
   final StreamController<LabelClickEvent> _onLabelClickController =
@@ -37,6 +28,10 @@ abstract class KakaoMapControllerPlatform extends PlatformInterface {
   final StreamController<CameraMoveEndEvent> _onCameraMoveEndController =
       StreamController<CameraMoveEndEvent>.broadcast();
 
+  /// Stream controller for cluster click events.
+  final StreamController<ClusterClickEvent> _onClusterClickController =
+      StreamController<ClusterClickEvent>.broadcast();
+
   /// Stream of label click events.
   Stream<LabelClickEvent> get onLabelClickedStream =>
       _onLabelClickController.stream;
@@ -49,13 +44,23 @@ abstract class KakaoMapControllerPlatform extends PlatformInterface {
   Stream<CameraMoveEndEvent> get onCameraMoveEndStream =>
       _onCameraMoveEndController.stream;
 
+  /// Stream of cluster click events.
+  Stream<ClusterClickEvent> get onClusterClickedStream =>
+      _onClusterClickController.stream;
+
   /// Calls a method on the platform implementation.
-  ///
-  /// Platform implementations must override this method to handle
-  /// method calls from the Dart side.
-  Future<T> _callMethod<T>(KakaoMapMethodCall<T> methodCall) async {
+  @internal
+  Future<T> callMethod<T>(KakaoMapMethodCall<T> methodCall) async {
     throw UnimplementedError(
       '[Flutter:KakaoMapControllerPlatform] callMethod not implemented',
+    );
+  }
+
+  /// Calls a web-specific method on the platform implementation.
+  @internal
+  Future<T> callWebMethod<T>(KakaoMapWebMethodCall<T> methodCall) async {
+    throw UnimplementedError(
+      '[Flutter:KakaoMapControllerPlatform] callWebMethod not implemented',
     );
   }
 
@@ -71,6 +76,10 @@ abstract class KakaoMapControllerPlatform extends PlatformInterface {
   void onCameraMoveEnd(CameraMoveEndEvent event) =>
       _onCameraMoveEndController.add(event);
 
+  /// Notifies listeners of a cluster click event.
+  void onClusterClicked(ClusterClickEvent event) =>
+      _onClusterClickController.add(event);
+
   /// Disposes of resources used by this platform interface.
   ///
   /// Closes all stream controllers to prevent memory leaks.
@@ -78,5 +87,6 @@ abstract class KakaoMapControllerPlatform extends PlatformInterface {
     _onLabelClickController.close();
     _onInfoWindowClickController.close();
     _onCameraMoveEndController.close();
+    _onClusterClickController.close();
   }
 }