+++ lib/src/data/map_widget/component/info_window_click_event.dart
@@ -33,7 +33,7 @@ class InfoWindowClickEvent extends Data {
 
   @override
   Map<String, Object?> toJson() => <String, Object?>{
-        'infoWindowId': infoWindowId,
-        'latLng': latLng.toJson(),
-      };
+    'infoWindowId': infoWindowId,
+    'latLng': latLng.toJson(),
+  };
 }