+++ lib/src/platform/platform.dart
@@ -6,6 +6,8 @@
 /// - 네이티브 Kakao Maps SDK와 통신하는 컨트롤러, 메서드 호출, 인터페이스 집합
 library kakao_maps_flutter.platform;
 
-export 'kakao_map_controller/kakao_map_controller.dart';
-export 'kakao_map_method_call/kakao_map_method_call.dart';
-export 'kakao_maps_flutter/kakao_maps_flutter.dart';
+export '../controller/kakao_map_controller.dart';
+export '../method_call/kakao_map_method_call.dart';
+export 'interface/kakao_maps_flutter_platform_interface.dart';
+export 'kakao_maps_flutter.dart';
+export 'method_channel/method_channel_kakao_maps_flutter.dart';